Skip over bad k8s auth.
Currently, if secret-updater is attempting to update something on a
cluster it cannot authenticate to (either bad auth, cluster removed, etc)
it fails fairly spectacularly, and fails to update all the clusters that
it can access.

Often times this failure (despite logged) can be missed, so the first
experience we have is secrets not getting updated when they should be.

Let's flip that about, lets skip over (but keep logging) those cases.